+++ OpenKit/OKDBScore.h
@@ -2,7 +2,7 @@
 //  OKDBScore.h
 //  OpenKit
 //
-//  Created by Suneet Shah on 7/26/13.
+//  Created by Manu Martinez-Almeida.
 //  Copyright (c) 2013 OpenKit. All rights reserved.
 //
 
@@ -16,7 +16,7 @@
 @property(nonatomic, strong) OKScore *previousSubmittedScore;
 
 - (NSArray*)getAllScores;
-- (NSArray*)getScoresForLeaderboardID:(int)leaderboardID
+- (NSArray*)getScoresForLeaderboardID:(NSInteger)leaderboardID
                         onlySubmitted:(BOOL)submittedOnly;
 - (void)clearSubmittedScores;
 - (NSArray*)getUnsubmittedScores;
