+++ OpenKit/OKHelper.h
@@ -13,7 +13,7 @@
 
 @interface OKHelper : NSObject
 
-+ (NSDate*)dateNDaysFromToday:(int)n;
++ (NSDate*)dateNDaysFromToday:(NSInteger)n;
 + (int)getIntFrom:(NSDictionary*)dict key:(NSString *)key;
 + (int64_t)getInt64From:(NSDictionary*)dict key:(NSString *)key;
 + (BOOL)getBOOLFrom:(NSDictionary*)dict key:(NSString *)key;
